sql_assessment_task_t-sql_query
DB Developer – SQL Assessment 
Consider that you want to identify clashes for a set of trainers based on their class schedule. Specifically, the class that starts while another class is still busy. Create a query to pull back all records that have overlapping datetimes for a specific Trainer from the sample rows below. In other words: clashes for each individual trainer, not clashes with another trainer’s classes. Your query should identify the clashing records as per the sample, but also for any other rows that can be added into the table in the future). Ensure your solution also work for new data that might be added into the table.